+++ example/go/service/file_service.go
@@ -9,7 +9,7 @@ import (
 	"math/rand"
 	"time"
 
-	pb "synurang/example/pkg/api"
+	pb "github.com/ivere27/synurang/example/pkg/api"
 	core_service "github.com/ivere27/synurang/pkg/service"
 
 	"google.golang.org/grpc"