+++ example/go/service/greeter.go
@@ -10,7 +10,7 @@ import (
 	"sync"
 	"time"
 
-	pb "synurang/example/pkg/api"
+	pb "github.com/ivere27/synurang/example/pkg/api"
 	core_service "github.com/ivere27/synurang/pkg/service"
 
 	"google.golang.org/grpc"